Veternas Page MVP completed
